+++ tiffin_web/src/components/header/headerRight/backButton/index.tsx
@@ -1,0 +1,9 @@
+import router from 'next/router'
+
+export const BackButton: React.FC = () => {
+  const goBack = () => {
+    router.back()
+  }
+
+  return <button onClick={goBack}>戻る</button>
+}
