+++ tiffin_web/src/components/header/headerRight/index.tsx
@@ -1,10 +1,9 @@
 import { ReactElement } from 'react'
-import styles from './headerRight.module.css'
 
 type HeaderRightProps = Required<{
   readonly children: ReactElement
 }>
 
 export const HeaderRight = ({ children }: HeaderRightProps) => {
-  return <div className={styles.headerRightArea}>{children}</div>
+  return <div>{children}</div>
 }
