--- tiffin_web/src/components/header/headerRight/signinButton/signinButton.tsx
@@ -1,14 +1,0 @@
-import router from 'next/router'
-import styles from './signinButton.module.css'
-
-export const SigninButton: React.FC = () => {
-  const pushSignin = () => {
-    router.push('/signin')
-  }
-
-  return (
-    <button className={styles.headerButton} onClick={pushSignin}>
-      signin
-    </button>
-  )
-}
