+++ tiffin_web/src/components/header/headerRight/signupButton/index.tsx
@@ -1,0 +1,9 @@
+import router from 'next/router'
+
+export const SignupButton: React.FC = () => {
+  const pushSignup = () => {
+    router.push('/signup')
+  }
+
+  return <button onClick={pushSignup}>signup</button>
+}
