+++ tiffin_web/src/components/header/index.tsx
@@ -1,5 +1,4 @@
 import { ReactElement } from 'react'
-import styles from './header.module.css'
 
 type HeaderProps = Required<{
   readonly children: ReactElement
@@ -8,7 +7,7 @@ type HeaderProps = Required<{
 export const Header = ({ children }: HeaderProps) => {
   return (
     <div>
-      <div className={styles.headerArea}>{children}</div>
+      <div>{children}</div>
     </div>
   )
 }
