+++ tiffin_web/src/components/layout/index.tsx
@@ -1,5 +1,4 @@
 import { ReactElement } from 'react'
-import styles from './layout.module.css'
 
 type LayoutProps = Required<{
   readonly children: ReactElement
@@ -8,7 +7,7 @@ type LayoutProps = Required<{
 export const Layout = ({ children }: LayoutProps) => {
   return (
     <div>
-      <div className={styles.container}>{children}</div>
+      <div>{children}</div>
     </div>
   )
 }
