+++ tiffin_web/src/pages/SignIn/index.tsx
@@ -2,12 +2,12 @@ import Axios from 'axios'
 import router from 'next/router'
 import { useState } from 'react'
 import styles from './signin.module.css'
-import { Header } from '@/components/header/header'
-import { AccountIcon } from '@/components/header/headerRight/accountIcon/accountIcon'
-import { HeaderRight } from '@/components/header/headerRight/headerRight'
-import { SignupButton } from '@/components/header/headerRight/signupButton/signupButton'
-import { Logo } from '@/components/header/logo/logo'
-import { Layout } from '@/components/layout/layout'
+import { Header } from '@/components/header'
+import { HeaderRight } from '@/components/header/headerRight'
+import { AccountIcon } from '@/components/header/headerRight/accountIcon'
+import { SignupButton } from '@/components/header/headerRight/signupButton'
+import { Logo } from '@/components/header/logo'
+import { Layout } from '@/components/layout'
 import { AuthSigninRequest } from '@/models/request/auth/authSigninRequest'
 import { AuthResponse } from '@/models/response/auth/authResponse'
 import { tiffin_token_key } from '@/utils/auth'
