+++ tiffin_web/src/pages/SignUp/index.tsx
@@ -2,12 +2,12 @@ import Axios from 'axios'
 import router from 'next/router'
 import { useState } from 'react'
 import styles from './signup.module.css'
-import { Header } from '@/components/header/header'
-import { AccountIcon } from '@/components/header/headerRight/accountIcon/accountIcon'
-import { HeaderRight } from '@/components/header/headerRight/headerRight'
-import { SigninButton } from '@/components/header/headerRight/signinButton/signinButton'
-import { Logo } from '@/components/header/logo/logo'
-import { Layout } from '@/components/layout/layout'
+import { Header } from '@/components/header'
+import { HeaderRight } from '@/components/header/headerRight'
+import { AccountIcon } from '@/components/header/headerRight/accountIcon'
+import { SigninButton } from '@/components/header/headerRight/signinButton'
+import { Logo } from '@/components/header/logo'
+import { Layout } from '@/components/layout'
 import { AuthSignupRequest } from '@/models/request/auth/authSignupRequest'
 
 export const Signup: React.FC = () => {
